Added movement to mini asteroids and began the linked list implementation
I've added movement to the mini asteroids similarly to how asteroids gete selected. Plus, i've began developing the linked list for storing the asteroids. Right now only have the initialization, the insert at the beggining, insert at end, and print of all the elements in the list. But i've ran into issues on using it on the mini asteroids script